Harden MPMC work publication
Track in-flight work transfers and generation changes so receivers never report premature disconnection while a batch is temporarily hidden.

Bound transient retries, publish ready lanes without a temporary Vec, and move sender and receiver code into focused modules. Add Loom and stress coverage for publication races.
