+++ tests/mpmc.rs
@@ -1,6 +1,8 @@
 use std::time::Instant;
 
-use fanring::mpmc::{RecvTimeoutError, SendTimeoutError, TrySendError, channel};
+use fanring::mpmc::{
+    RecvError, RecvTimeoutError, SendTimeoutError, TryRecvError, TrySendError, channel,
+};
 
 #[test]
 fn cloned_receivers_distribute_messages() {
@@ -38,7 +40,7 @@ fn receiver_drop_preserves_prefetched_work() {
     tx.try_send(2).unwrap();
     assert_eq!(rx0.try_recv(), Ok(1));
     drop(rx0);
-    assert_eq!(rx1.try_recv(), Ok(2));
+    assert_eq!(rx1.recv(), Ok(2));
 }
 
 #[test]
@@ -140,4 +142,75 @@ fn receiver_iterators_cover_blocking_and_ready_values() {
     tx.try_send(5).unwrap();
     drop(tx);
     assert_eq!(rx.into_iter().collect::<Vec<_>>(), [5]);
+}
+
+#[test]
+fn receiver_reports_sender_disconnect_before_buffer_is_drained() {
+    let (mut tx, mut rx) = channel(2);
+    tx.try_send(1).unwrap();
+    tx.try_send(2).unwrap();
+    drop(tx);
+
+    assert!(rx.is_disconnected());
+    assert_eq!(rx.recv(), Ok(1));
+    assert_eq!(rx.recv(), Ok(2));
+    assert_eq!(rx.recv(), Err(RecvError));
+}
+
+#[test]
+fn blocking_iterators_stay_exhausted_after_disconnect() {
+    let (mut tx, mut rx) = channel(1);
+    tx.try_send(1).unwrap();
+    drop(tx);
+
+    let mut iter = rx.iter();
+    assert_eq!(iter.next(), Some(1));
+    assert_eq!(iter.next(), None);
+    assert_eq!(iter.next(), None);
+}
+
+#[test]
+fn competing_receiver_may_observe_transient_empty() {
+    let (mut tx, mut rx0) = channel(64);
+    let mut rx1 = rx0.clone();
+    for value in 0..64 {
+        tx.try_send(value).unwrap();
+    }
+
+    assert_eq!(rx0.try_recv(), Ok(0));
+    assert!(matches!(rx1.try_recv(), Ok(_) | Err(TryRecvError::Empty)));
+}
+
+#[test]
+fn zero_sized_and_highly_aligned_values_round_trip() {
+    #[derive(Debug, PartialEq, Eq)]
+    struct ZeroSized;
+
+    #[repr(align(256))]
+    #[derive(Debug, PartialEq, Eq)]
+    struct Aligned(usize);
+
+    let (mut tx, mut rx) = channel(2);
+    tx.try_send(ZeroSized).unwrap();
+    tx.try_send(ZeroSized).unwrap();
+    assert_eq!(rx.try_recv(), Ok(ZeroSized));
+    assert_eq!(rx.try_recv(), Ok(ZeroSized));
+
+    let (mut tx, mut rx) = channel(2);
+    tx.try_send(Aligned(1)).unwrap();
+    tx.try_send(Aligned(2)).unwrap();
+    assert_eq!(rx.try_recv(), Ok(Aligned(1)));
+    assert_eq!(rx.try_recv(), Ok(Aligned(2)));
+}
+
+#[test]
+fn non_copy_values_survive_receiver_handoff() {
+    let (mut tx, mut rx0) = channel(4);
+    tx.try_send(String::from("first")).unwrap();
+    tx.try_send(String::from("second")).unwrap();
+    assert_eq!(rx0.try_recv().as_deref(), Ok("first"));
+
+    let mut rx1 = rx0.clone();
+    drop(rx0);
+    assert_eq!(rx1.recv().as_deref(), Ok("second"));
 }
